blue2yellow {colorRamps}	R Documentation

Returns a gradient color map

Description

blue2yellow makes a blue to yellow gradient color map

Usage

blue2yellow(n)
cyan2yellow(n)
magenta2green(n)

Arguments

n
number of colors

Details

These are single gradient maps that smoothly transition from cooler to warmer colors. See blue2red for double gradient maps.

Value

A color map

Author(s)

Tim Keitt <tkeitt@gmail.com>

References

Keitt, T. H. (2008) Coherent ecological dynamics induced by large scale disturbance. Nature 454:331-334

See Also

rgb

Examples

image(matrix(1:400, 20), col = blue2yellow(400))

[Package colorRamps version 2.3 Index]