colorRamps-package {colorRamps}	R Documentation

Builds color maps

Description

This (v2) is a rewrite of the colorRamps package. It now contains two function table.ramp and rgb.tables that allow easy construction of color palettes. This version contains two new palettes similar to the Matlab default palette (matlab.like and matlab.like2).

I built colorRamps because I needed to use a particular palette and got tired of sourcing in my code into every session. Now I can install and forget. Despite using R for years, I had not noticed the alternative colorRamp which may suit your needs. If you want really attractive palettes, get the RColorBrewer package from CRAN. For certain applications the RColorBrewer palettes do not work for me, hence this package.

Details

Package: 	 colorRamps

Type: 	 Package

Version: 	 2.0

Date: 	 2007-09-09

License: 	 GPL

Most functions take a single argument n that specifies the number of colors to generate.

Author(s)

Tim Keitt

Maintainer: Tim Keitt <tkeitt@gmail.com>

References

Keitt, T. H. (2008) Coherent ecological dynamics induced by large scale disturbance. Nature 454:331-334

Examples

filled.contour(volcano, col = ygobb(21), asp = 1)

[Package colorRamps version 2.3 Index]